1.0.0 alpha

+++ bot/bot.py
@@ -8,24 +8,28 @@ from discord.ext import commands as cmd
 from . import models
 from .utils import Utils
 
+client = AsyncIOMotorClient(mongo)
+
 
 class Ciri(cmd.Bot):
     def __init__(self):
         super().__init__(command_prefix=self.get_prefix, case_insensitive=True, intents=discord.Intents.all())
-        self.client = AsyncIOMotorClient(mongo)
+        self.client = client
 
     def load(self):
         self.models = models
         self.utils = Utils(self)
         self.db = self.client
-        self.config = self.client.get_database("config").get_collection("main")
+        self.config = self.client.get_database("cfg").get_collection("main")
         self.profiles = self.client.get_database("users").get_collection("profiles")
         self.server = self.client.get_database("server")
         self.db_enumeration = {
             "main": 0,
             "profile": 1,
             "rep": 2,
-            "add_rep": 3
+            "balance": 3,
+            "timely": 4,
+            "pay": 5
         }
         self.remove_command('help')
 
@@ -39,7 +43,7 @@ class Ciri(cmd.Bot):
         print('-' * 30)
 
     async def on_command(self, ctx: cmd.Context):
-        if not ctx.author.bot and ctx.message.guild:
+        if not ctx.author.bot and ctx.message.guild and ctx.command.cog.qualified_name != "Arts":
             try:
                 await ctx.message.delete()
             except:
@@ -47,14 +51,14 @@ class Ciri(cmd.Bot):
 
     async def get_prefix(self, message: discord.Message):
         if not message.author.bot:
-            return "+"
+            return "t+"
 
-    async def on_command_error(self, ctx: cmd.Context, err):
-        if not ctx.command:
-            return print(err)
-        embed = discord.Embed(title=ctx.command.name + " error", description=str(err))
-        embed.set_footer(text=ctx.command.usage)
-        await ctx.send(embed=embed)
+    # async def on_command_error(self, ctx: cmd.Context, err):
+    #     if not ctx.command:
+    #         return print(err)
+    #     embed = discord.Embed(title=ctx.command.name + " error", description=str(err))
+    #     embed.set_footer(text=ctx.command.usage)
+    #     await ctx.send(embed=embed)
 
     async def on_ready(self):
         self.load()
